[Core] Make UpdateCurrentValue awaitable
This makes UpdateCurrentValue into +Async and return a task for
awaiting. Although the error handling behavior was intact due to the UI
threaded nature of VMs, there were cases where event signaling could get
out of order by calling this and then raising the change notification
without waiting (because you couldn't.) There are cases where we can
safely fire off a request for the value to be updated, so now these
cases are distinct with the addition of RequestValueUpdate() which is
explicitly not a sync update, but a request.